Fix createsuperuser/changepassword
The issue comes from the User, InteractiveUser and TechnicalUser all inheriting from the base AbstractUser and having a _u delegation. This means that Django cannot guess from User that it has to request a password when using createsuperuser. So we're allowing an empty password by default.
This obviously requires to set the password in a second command. The changepassword was not working either because when setting the password, it was properly delegated to the underlying TechnicalUser but that user was not saved, only User. I therefore added a save() function to delegate the saving of the underlying object if it had already been created (it has an id).