OMT-215 Add APscheduler
The regular way to enable the schedule in Django is to call a method from the Config.ready(). However, this is not very flexible and doesn't work with our modular architecture. So the core takes a couple of lists from the settings and dynamically invokes the corresponding methods.
There are two configuration lists: jobs to be added and customer scheduler methods if a simple add_job is not enough
The @shared_task annotation is optional.