Add decorator to translate queries
Mutations inherit from OpenIMISMutation that calls the subclasses' async_mutate and therefore can easily switch the language according to the logged user's preference.
For queries, this is not so easy since one can define any resolve_* method that will be called by GraphQL. To work around this, I added a decorator @translate_query that can be added to resolve_* methods that might need to translate error messages.
Note that it currently only works for positional arguments, I don't think that this would be a problem but otherwise, just update the for else with a second loop, if kwargs is defined.