Link User to ClaimAdmin and Officer + Queries
The users query is now querying over the ClaimAdmin and Officer types.

This is not really a clean solution but it is the best that we could make with the existing table structures without breaking everything else.
Be careful not to optimize the graphql query or it would prevent the User dynamic mapping to the actual objects because some of them would be marked as "deferred".
The Mutations are not yet part of this commit but will follow shortly.