Missing UserMutationGQLType + Mutation by ID
The mutation link to actual objects was performed on the actual object itself. In the case of User, it caused some conflicts with its `__call__` method because _u was not properly initialised from Django itself. So we are now adding _id to the object and passing its id. We still need the object itself to identify the foreign key.
UserMutation was also not exposed to GraphQL and the user was not properly linked for the mutation, it was linked to connected user.