Security fix for Django order_by flaw
Details: https://github.com/advisories/GHSA-xpfp-f569-q3p2

The proper way to address this would be to upgrade Django to 3.1.13 but it requires to update all the JSONField (and convert the NullBooleanField to BooleanField). While the JSONField seems to work when we remove the JSONFallback component, the boolean field does not work anymore with `bit` data types. They are compared with the boolean instead of 0 or 1.
I tried to check for updates on the mssql driver but it seems not to have moved in a long while.

I have therefore implemented the same kind of fix as the 3.1.13 did, right into the OrderedDjangoConnectionField...