Cannot connect with non-admin user OMT-281
This is not exactly OMT-281 but being unable to connect with a claim administrator prevents the test at all.
As pointed out in the comment, by default, the rest framework will look for core.user_view permission among the numeric permissions of openIMIS and obviously deny access. Explicitly setting the permission on that specific view allows any authenticated user to view their own data, as intended. Unauthenticated users will still get a 403 response.