calc rule calculate() should not name first arg
The method was declared as `instance, *args` so calling calculate(instance=instance, *args) failed with unexpected parameter signal (first kwargs). Calling the method with instance=instance makes it a named argument, so you can't pass any positional argument after that. That proper way is for methods to accept both *args and **kwargs and then call with instance as positional, the *args will add more positional arguments and then **kwargs can be passed.

This makes the contract module test properly.