OMT-323 - Registers status report (#207)
* v1.2.1

* Missing __init__ in gql folder for setup

* v1.3.0

* Cannot connect with non-admin user OMT-281

This is not exactly OMT-281 but being unable to connect with a claim administrator prevents the test at all.
As pointed out in the comment, by default, the rest framework will look for core.user_view permission among the numeric permissions of openIMIS and obviously deny access. Explicitly setting the permission on that specific view allows any authenticated user to view their own data, as intended. Unauthenticated users will still get a 403 response.

* fix(Performance): Use legacy_id instead of validity_to to filter historical data
MS SQL Server is super slow and does not use correctly the indexes when using a filter with a date

* feat(Sentry): Better logging of errors in mutations

* hotfix(User): Keep user roles history

* Facilitate access to health facility

The interactive user has a health facility without foreign key but the claim admin has a foreign key. This method encapsulates this and is needed for OMT-281 so that price lists are available.

* v1.3.1

* OMT-281 follow-up when HFID is 0

The cached_property would fail when called with HFID==0 (as happens in tblUsers that has no foreign key on this one)

* v.1.3.3

* Add Migration on JSONExt column already fixed on this PR : https://github.com/openimis/openimis-be-core_py/pull/186
Need to be merged as an hotfix to solve the policy issue on main release

* Hotfix is_imis_admin missing from release

* OMT-323 - Registers status report

---------

Co-authored-by: xgill <38523490+xgill@users.noreply.github.com>
Co-authored-by: Eric Darchis <edarchis@bluesquarehub.com>
Co-authored-by: Eric Darchis <darchis@pobox.com>
Co-authored-by: Patrick Delcroix <patrick.delcroix@swisstph.ch>
Co-authored-by: Dragos DOBRE <dragos.dobre@swisstph.ch>
Co-authored-by: Quentin Gérôme <geromequentin@gmail.com>
Co-authored-by: Hiren Soni <65155846+hirensoni913@users.noreply.github.com>
Co-authored-by: Maxime NGOE <maxime.ngoe@gmail.com>
Co-authored-by: Kamil Malinowski <kmalinowski@soldevelo.com>